+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # BuildAssist
 
+## [1.0.1] - 2020-03-23
+
+### Added
+- Added "PlayerSettings.WebGL.compressionFormat" to diff
+
+### Fixed
+- Fixed error when opening preferences
+
 ## [1.0.0] - 2020-03-21
 - First release